+++ packages/pytest-simcore/src/pytest_simcore/httpx_calls_capture.py
@@ -1,14 +1,26 @@
+# pylint: disable=protected-access
 # pylint: disable=redefined-outer-name
+# pylint: disable=too-many-arguments
 # pylint: disable=unused-argument
 # pylint: disable=unused-variable
-# pylint: disable=too-many-arguments
 
+import json
 from collections.abc import Callable
 from pathlib import Path
+from typing import Any
 
+import httpx
 import pytest
+import respx
+from pydantic import parse_obj_as
 from pytest_mock import MockerFixture, MockType
 from pytest_simcore.helpers.httpx_client_base_dev import AsyncClientCaptureWrapper
+
+from .helpers.httpx_calls_capture_model import (
+    HttpApiCallCaptureModel,
+    PathDescription,
+    SideEffectCallback,
+)
 
 
 def pytest_addoption(parser: pytest.Parser):
@@ -62,3 +74,97 @@ def spy_async_client_or_none(
         return None
 
     return _
+
+
+class _CaptureSideEffect:
+    def __init__(
+        self,
+        capture: HttpApiCallCaptureModel,
+        side_effect: SideEffectCallback | None,
+    ):
+        self._capture = capture
+        self._side_effect_callback = side_effect
+
+    def __call__(self, request: httpx.Request, **kwargs) -> httpx.Response:
+        capture = self._capture
+        assert isinstance(capture.path, PathDescription)
+        status_code: int = capture.status_code
+        response_body: dict[str, Any] | list | None = capture.response_body
+        assert {param.name for param in capture.path.path_parameters} == set(
+            kwargs.keys()
+        )
+        if self._side_effect_callback:
+            response_body = self._side_effect_callback(request, kwargs, capture)
+        return httpx.Response(status_code=status_code, json=response_body)
+
+
+@pytest.fixture
+@respx.mock(assert_all_mocked=False)
+def respx_mock_from_capture() -> (
+    Callable[
+        [list[respx.MockRouter], Path, list[SideEffectCallback]], list[respx.MockRouter]
+    ]
+):
+    def _generate_mock(
+        respx_mock: list[respx.MockRouter],
+        capture_path: Path,
+        side_effects_callbacks: list[SideEffectCallback],
+    ) -> list[respx.MockRouter]:
+        assert capture_path.is_file()
+        assert capture_path.suffix == ".json"
+
+        captures: list[HttpApiCallCaptureModel] = parse_obj_as(
+            list[HttpApiCallCaptureModel], json.loads(capture_path.read_text())
+        )
+
+        if len(side_effects_callbacks) > 0:
+            assert len(side_effects_callbacks) == len(captures)
+
+        assert isinstance(respx_mock, list)
+        for router in respx_mock:
+            assert (
+                router._bases
+            ), "the base_url must be set before the fixture is extended"
+
+        def _get_correct_mock_router_for_capture(
+            respx_mock: list[respx.MockRouter], capture: HttpApiCallCaptureModel
+        ) -> respx.MockRouter:
+            for router in respx_mock:
+                if capture.host == router._bases["host"].value:
+                    return router
+            msg = f"Missing respx.MockRouter for capture with {capture.host}"
+            raise RuntimeError(msg)
+
+        side_effects: list[_CaptureSideEffect] = []
+        for ii, capture in enumerate(captures):
+            url_path: PathDescription | str = capture.path
+            assert isinstance(url_path, PathDescription)
+
+            # path
+            path_regex: str = str(url_path.path)
+            for param in url_path.path_parameters:
+                path_regex = path_regex.replace(
+                    "{" + param.name + "}", param.respx_lookup
+                )
+
+            # response
+            side_effect = _CaptureSideEffect(
+                capture=capture,
+                side_effect=side_effects_callbacks[ii]
+                if len(side_effects_callbacks)
+                else None,
+            )
+
+            router = _get_correct_mock_router_for_capture(respx_mock, capture)
+            r = router.request(
+                capture.method.upper(),
+                url=None,
+                path__regex=f"^{path_regex}$",
+            ).mock(side_effect=side_effect)
+
+            assert r.side_effect == side_effect
+            side_effects.append(side_effect)
+
+        return respx_mock
+
+    return _generate_mock
