+++ services/api-server/tests/unit/test_api_solvers.py
@@ -1,3 +1,8 @@
+# pylint: disable=redefined-outer-name
+# pylint: disable=unused-argument
+# pylint: disable=unused-variable
+# pylint: disable=too-many-arguments
+
 from collections.abc import Callable
 from pathlib import Path
 
@@ -8,8 +13,8 @@ from fastapi import status
 from httpx import AsyncClient
 from models_library.api_schemas_api_server.pricing_plans import ServicePricingPlanGet
 from pydantic import parse_obj_as
+from pytest_simcore.helpers.httpx_calls_capture_model import SideEffectCallback
 from simcore_service_api_server._meta import API_VTAG
-from unit.conftest import SideEffectCallback
 
 
 @pytest.mark.parametrize(
