+++ services/api-server/tests/unit/test_api_wallets.py
@@ -1,3 +1,8 @@
+# pylint: disable=redefined-outer-name
+# pylint: disable=unused-argument
+# pylint: disable=unused-variable
+# pylint: disable=too-many-arguments
+
 from collections.abc import Callable
 from pathlib import Path
 from typing import Any
@@ -9,12 +14,11 @@ from fastapi import status
 from httpx import AsyncClient
 from models_library.api_schemas_webserver.wallets import WalletGetWithAvailableCredits
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_model import (
+    HttpApiCallCaptureModel,
+    SideEffectCallback,
+)
 from simcore_service_api_server._meta import API_VTAG
-from unit.conftest import SideEffectCallback
-
-# pylint: disable=unused-argument
-# pylint: disable=unused-variable
 
 
 @pytest.mark.parametrize(
